requirements.txt
langchain
faiss-cpu   
openai 
langchain-openai
python-dotenv
langchain-community